git-svn-id: svn+ssh://src.apple.com/svn/cups/easysw/current@624 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ scheduler/client.c
@@ -2530,7 +2530,7 @@ cupsdSendHeader(
   {
     cupsdLogMessage(CUPSD_LOG_ERROR,
                     "Kerberos credentials larger than 64k (%d)!",
-		    con->gss_output_token.length);
+		    (int)con->gss_output_token.length);
     return (0);
   }
 #endif /* HAVE_GSSAPI */
